Refactor Android and JS rendering, update dependencies.
Adjusted rendering logic for better integration with native APIs and improved type safety by switching to unsigned integers where needed. Updated dependency declarations and corrected some inconsistencies in texture format handling across platforms.

+++ wgpu4k-scenes/src/wasmJsMain/kotlin/io.ygdrasil.wgpu.examples/AssetManager.wasmJs.kt
@@ -10,5 +10,5 @@ actual var customVfs: Vfs
     set(value) {}
 
 internal actual fun Bitmap32.toBitmapHolder(textureFormat: ColorFormat): ImageBitmapHolder {
-    return ImageBitmapHolder(width, height, extractBytes(textureFormat))
+    return ImageBitmapHolder(width.toUInt(), height.toUInt(), extractBytes(textureFormat))
 }
